Simplify chat to use current VivaMode provider/model
Remove per-conversation provider/model selection. Chat now uses
whatever provider/model the active VivaMode has. Removes
ChatProviderPickerView, aiProviderName/aiModelName from
ChatConversation, and updateProvider/updateModel methods.
Header bar shows current provider + model as read-only label.
